+++ packages/tests/src/main.ts
@@ -1,16 +1,19 @@
-import { PropertyValues, state } from "snar";
-import { ReactiveLocalStorage } from "snar/lib/local-storage.js";
+import {state} from 'snar';
+import {ReactiveLocalStorage} from 'snar/lib/localstorage.js';
+import {SingleHostController} from '@snar/lit';
+
+console.log(SingleHostController)
 
 class LS1 extends ReactiveLocalStorage {
-  @state() foo = "foo";
+	@state() foo = 'foo';
 }
 class LS2 extends ReactiveLocalStorage {
-  @state() bar = "bar";
+	@state() bar = 'bar';
 }
 
-await new LS1("same-handler").updateComplete;
+await new LS1('same-handler').updateComplete;
 try {
-  await new LS2("same-handler").updateComplete;
+	await new LS2('same-handler').updateComplete;
 } catch (e) {
-  console.log("lol");
+	console.log('lol');
 }